Align with Jetpack Changes (toBundle/FW Flag)
Recently, we added a change to Jetpack to bundle BPData properties
within a bundle, rather than directly imbue it into the slice.
This requires us to re-structure our slice parsing, to:

1. Read the bundle from the slice.
2. Parse the bundle after reading from the slice.

Along this change, we also need to know, for a whole many use cases,
when the framework prompt was the originator for the codes of the BP.
This is also a Jetpack decision, so a small bit was added so it can be
utilized in JP.

Bug: 339650358
Test: Build and device tests

+++ packages/CredentialManager/src/com/android/credentialmanager/common/BiometricHandler.kt
@@ -230,7 +230,7 @@ private fun runBiometricFlow(
         val cryptoOpId = getCryptoOpId(biometricDisplayInfo)
         if (cryptoOpId != null) {
             biometricPrompt.authenticate(
-                BiometricPrompt.CryptoObject(cryptoOpId.toLong()),
+                BiometricPrompt.CryptoObject(cryptoOpId),
                 cancellationSignal, executor, callback)
         } else {
             biometricPrompt.authenticate(cancellationSignal, executor, callback)
@@ -243,7 +243,7 @@ private fun runBiometricFlow(
     return true
 }
 
-private fun getCryptoOpId(biometricDisplayInfo: BiometricDisplayInfo): Int? {
+private fun getCryptoOpId(biometricDisplayInfo: BiometricDisplayInfo): Long? {
     return biometricDisplayInfo.biometricRequestInfo.opId
 }
